Data-driven animation loading

+++ src/overworld/render-map.cpp
@@ -13,7 +13,6 @@ void renderMap(
     engine::resourcesystem::ResourceStorage& storage
 ) {
     using engine::entitysystem::Entity;
-    // using namespace engine::spritesystem;
 
     manager.forEachEntity<Map>(
         [&](
@@ -21,8 +20,8 @@ void renderMap(
             Map& map
         ) {
             for (size_t i = 0; i < map.tiles.size(); ++i) {
-                int x = (i % map.widthInTiles) * TILE_SIZE;
-                int y = (i / map.widthInTiles) * TILE_SIZE;
+                float x = (i % map.widthInTiles) * TILE_SIZE;
+                float y = (i / map.widthInTiles) * TILE_SIZE;
                 Tile& tile = map.tiles.at(i);
                 tile.sprite.setPosition({x, y});
                 window.draw(tile.sprite);
